requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
pymongo==4.14.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.179.0
pytesseract==0.3.13
Pillow==11.3.0
opencv-python==4.9.0.80
numpy==1.26.4
pydantic==2.11.7
